fix(library): keep dev perf counters in a leaf module so yjsSync stays its own chunk
The DEV-only perf counters lived in sync/yjsSync, so importing them from
diagramStore and apollon-editor created a value edge into yjsSync from the
eager graph and merged it into the main bundle — removing the separate
yjsSync chunk that size-limit budgets. Move the counters to a dependency-free
sync/perfCounters leaf module so consumers no longer pull yjsSync in, restoring
the chunk. Production cost is unchanged (the module dead-code-eliminates under
import.meta.env.DEV; the leak guard still passes).
