+++ library/tests/unit/perfProbe.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from "vitest"
 import * as Y from "yjs"
 import { ApollonEditor } from "@/apollon-editor"
 import { createDiagramStore } from "@/store/diagramStore"
-import { getPerfCounters } from "@/sync/yjsSync"
+import { getPerfCounters } from "@/sync/perfCounters"
 import { getNodesMap, STORE_ORIGIN } from "@/sync/ydoc"
 import { UMLDiagramType } from "@/types"
 import type { UMLModel } from "@/typings"
